+++ adept/sh2d/solvers/tridiagonal.py
